+++ client/rpc-api/Cargo.toml
@@ -23,6 +23,7 @@ log = "0.4.8"
 parking_lot = "0.12.0"
 thiserror = "1.0"
 
+scale-info = { version = "2.0.1", default-features = false, features = ["derive"] }
 sp-core = { version = "6.0.0", path = "../../primitives/core" }
 sp-version = { version = "5.0.0", path = "../../primitives/version" }
 sp-runtime = { version = "6.0.0", path = "../../primitives/runtime" }
